worktree/minsu/pill-detection-project/requirements.txt
# requirements.txt
# 알약 검출 프로젝트 의존성

# 딥러닝 프레임워크
torch>=1.10.0
torchvision>=0.11.0

# 데이터 처리
numpy>=1.19.0
pandas>=1.3.0
Pillow>=8.0.0
opencv-python>=4.5.0

# 데이터 증강
albumentations>=1.0.0

# 설정 관리
PyYAML>=5.4.0

# 시각화
matplotlib>=3.3.0
seaborn>=0.11.0
tensorboard>=2.7.0

# 유틸리티
tqdm>=4.62.0
scikit-learn>=0.24.0
ultralytics

# 개발 도구 (선택사항)
jupyter>=1.0.0
ipywidgets>=7.6.0
